+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/js/Obj.kt
@@ -62,3 +62,22 @@ internal data class AdCallbackObj(val param: AdObj)
 internal const val SUCCESS = "success"
 internal const val CLOSED = "closed"
 // end region
+
+// send contact region
+@Keep
+internal data class SendContactCallbackObj(val param: MessageParam) {
+
+    @Keep
+    internal data class MessageParam(val messageToContact: MessageToContact)
+}
+
+@Keep
+internal data class SendContactIdCallbackObj(val param: MessageParamId) {
+
+    @Keep
+    internal data class MessageParamId(
+        val contactId: String,
+        val messageToContact: MessageToContact
+    )
+}
+// end region