+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/chat/SelectableContact.kt
@@ -1,0 +1,10 @@
+package com.rakuten.tech.mobile.testapp.ui.chat
+
+import androidx.annotation.Keep
+import com.rakuten.tech.mobile.miniapp.js.userinfo.Contact
+
+@Keep
+data class SelectableContact(
+    val contact: Contact,
+    var isSelected: Boolean = false
+)