+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/display/MiniAppDisplayActivity.kt
@@ -17,10 +17,10 @@ import androidx.lifecycle.ViewModelProvider
 import com.rakuten.tech.mobile.miniapp.MiniAppInfo
 import com.rakuten.tech.mobile.miniapp.ads.AdMobDisplayer
 import com.rakuten.tech.mobile.miniapp.file.MiniAppFileChooserDefault
+import com.rakuten.tech.mobile.miniapp.js.MessageToContact
 import com.rakuten.tech.mobile.miniapp.js.MiniAppMessageBridge
-import com.rakuten.tech.mobile.miniapp.js.userinfo.Contact
-import com.rakuten.tech.mobile.miniapp.js.userinfo.TokenData
-import com.rakuten.tech.mobile.miniapp.js.userinfo.UserInfoBridgeDispatcher
+import com.rakuten.tech.mobile.miniapp.js.chat.ChatBridgeDispatcher
+import com.rakuten.tech.mobile.miniapp.js.userinfo.*
 import com.rakuten.tech.mobile.miniapp.navigator.ExternalResultHandler
 import com.rakuten.tech.mobile.miniapp.navigator.MiniAppNavigator
 import com.rakuten.tech.mobile.miniapp.permission.AccessTokenScope
@@ -29,6 +29,7 @@ import com.rakuten.tech.mobile.miniapp.testapp.R
 import com.rakuten.tech.mobile.miniapp.testapp.databinding.MiniAppDisplayActivityBinding
 import com.rakuten.tech.mobile.testapp.helper.AppPermission
 import com.rakuten.tech.mobile.testapp.ui.base.BaseActivity
+import com.rakuten.tech.mobile.testapp.ui.chat.ChatWindow
 import com.rakuten.tech.mobile.testapp.ui.settings.AppSettings
 import java.util.*
 
@@ -146,6 +147,7 @@ class MiniAppDisplayActivity : BaseActivity() {
     }
 
     private fun setupMiniAppMessageBridge() {
+        // setup MiniAppMessageBridge
         miniAppMessageBridge = object : MiniAppMessageBridge() {
             override fun getUniqueId() = AppSettings.instance.uniqueId
 
@@ -165,6 +167,7 @@ class MiniAppDisplayActivity : BaseActivity() {
         miniAppMessageBridge.setAdMobDisplayer(AdMobDisplayer(this@MiniAppDisplayActivity))
         miniAppMessageBridge.allowScreenOrientation(true)
 
+        // setup UserInfoBridgeDispatcher
         val userInfoBridgeDispatcher = object : UserInfoBridgeDispatcher {
 
             override fun getUserName(
@@ -203,6 +206,37 @@ class MiniAppDisplayActivity : BaseActivity() {
             }
         }
         miniAppMessageBridge.setUserInfoBridgeDispatcher(userInfoBridgeDispatcher)
+
+        // setup ChatBridgeDispatcher
+        val chatWindow = ChatWindow(this@MiniAppDisplayActivity)
+        val chatBridgeDispatcher = object : ChatBridgeDispatcher {
+
+            override fun sendMessageToContact(
+                message: MessageToContact,
+                onSuccess: (contactId: String?) -> Unit,
+                onError: (message: String) -> Unit
+            ) {
+                chatWindow.openSingleContactSelection(message, onSuccess, onError)
+            }
+
+            override fun sendMessageToContactId(
+                contactId: String,
+                message: MessageToContact,
+                onSuccess: (contactId: String?) -> Unit,
+                onError: (message: String) -> Unit
+            ) {
+                chatWindow.openSpecificContactIdSelection(contactId, message, onSuccess, onError)
+            }
+
+            override fun sendMessageToMultipleContacts(
+                message: MessageToContact,
+                onSuccess: (contactIds: List<String>?) -> Unit,
+                onError: (message: String) -> Unit
+            ) {
+                chatWindow.openMultipleContactSelections(message, onSuccess, onError)
+            }
+        }
+        miniAppMessageBridge.setChatBridgeDispatcher(chatBridgeDispatcher)
     }
 
     override fun onRequestPermissionsResult(